requirements.txt
pytorch
tensorboardX
numpy
